+++ 40col-color/picoterm_core.c
@@ -813,7 +813,7 @@ void handle_new_character(unsigned char asc){
           if(insert_mode) insert_chars(1);
 
           // --- Strict ASCII <0x7f or Extended NuPetSCII <= 0xFF ---
-          slip_character(asc-32,conio_config.cursor.pos.x,conio_config.cursor.pos.y);
+          put_char(asc-32,conio_config.cursor.pos.x,conio_config.cursor.pos.y);
           conio_config.cursor.pos.x++;
 
           if(!conio_config.wrap_text){