+++ common/picoterm_stdio.h
@@ -1,0 +1,13 @@
+/* ==========================================================================
+    Manage the adcanced interaction with the console (printing string,
+		request input).
+
+		Relies on the Picoterm_cooio.c to interact with screen
+   ========================================================================== */
+
+#ifndef _PICOTERM_STDIO_H_
+#define _PICOTERM_STDIO_H_
+
+void print_string(char str[] );
+
+#endif // _PICOTERM_STDIO_H_